Replace usage of Queue for thread communication.
This patch uses the Event pattern described in
https://emptysqua.re/blog/an-event-synchronization-primitive-for-ruby/
to force the tests to wait for the WEBrick server to start successfully
before beginning any tests.  Previously, the thread running the Serve
class was pushing some data into a Queue to indicate it had started
while the test thread slept and woke periodically to check the Queue.